Split generic dialogs into component
