+++ src/dialogs/ConfirmDialog.tsx
@@ -1,11 +1,7 @@
-import React, {useState} from 'react';
-import {StyleSheet, View} from 'react-native';
-import colors from '../styles/colors';
-import {Dialog, Button} from '@rneui/themed';
-import {RegularText} from '../components/RegularText';
-import {styles} from '../styles/styles';
+import React from 'react';
+import {GenericDialog} from './GenericDialog';
 import {ActionCb} from '../types/ActionCb';
-import {NoDialogButton, YesDialogButton} from '../components/DialogButtons';
+import {View} from 'react-native';
 
 export interface ConfirmDialogProps {
   action: ActionCb;
@@ -26,44 +22,16 @@ export interface ConfirmDialogProps {
  *   action={() => console.log('action')} />
  */
 export function ConfirmDialog(props: ConfirmDialogProps): JSX.Element {
-  const [visible, setVisible] = useState(false);
-  const toggleDialog = () => setVisible(!visible);
-  const buttonPressed = () => toggleDialog();
-
-  const yesPressed = () => {
-    props.action();
-    toggleDialog();
-  };
+  const yesPressed = () => props.action();
 
   return (
     <View>
-      <Button
-        buttonStyle={styles.buttonOpenDialogStyle}
-        onPress={buttonPressed}>
-        <RegularText>{props.openBtnTitle}</RegularText>
-      </Button>
-      <Dialog
-        overlayStyle={deleteCommentDialogStyles.container}
-        style={deleteCommentDialogStyles.container}
-        isVisible={visible}
-        onBackdropPress={toggleDialog}>
-        <Dialog.Title titleStyle={styles.titleTextDialog} title={props.title} />
-        <RegularText>{props.message}</RegularText>
-        <Dialog.Actions>
-          <View style={styles.containerButtonDialog}>
-            <NoDialogButton onPress={() => setVisible(false)} />
-            <YesDialogButton onPress={yesPressed} />
-          </View>
-        </Dialog.Actions>
-      </Dialog>
+      <GenericDialog
+        openBtnTitle={props.openBtnTitle}
+        title={props.title}
+        message={props.message}
+        action={yesPressed}
+      />
     </View>
   );
 }
-
-const deleteCommentDialogStyles = StyleSheet.create({
-  container: {
-    paddingHorizontal: 20,
-    backgroundColor: colors.mediumBlack,
-    paddingVertical: 12,
-  },
-});
